consolidate user model and queries
