--- backend/src/models/user.ts
@@ -1,5 +1,0 @@
-export interface Teacher {
-  id: string;
-  email: string;
-  password_hash: string;
-}
